+++ main.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"net"
 	"net/http"
-	"os"
 
 	"github.com/buildwithgrove/path/envoy/auth_server/proto"
 	"github.com/pokt-network/poktroll/pkg/polylog/polyzero"
@@ -72,38 +71,4 @@ func main() {
 	if err := httpServer.Serve(ln); err != nil {
 		panic(fmt.Sprintf("failed to serve: %v", err))
 	}
-}
-
-/* ---------------------- Environment Variables ---------------------- */
-
-const (
-	yamlFilePathEnv = "YAML_FILEPATH"
-
-	portEnv     = "PORT"
-	defaultPort = "50051"
-)
-
-type envVars struct {
-	yamlFilepath string
-	port         string
-}
-
-func gatherEnvVars() (envVars, error) {
-	env := envVars{
-		yamlFilepath: os.Getenv(yamlFilePathEnv),
-		port:         os.Getenv(portEnv),
-	}
-	return env, env.validateAndHydrate()
-}
-
-// validateAndHydrate validates the required environment variables are set
-// and hydrates defaults for any optional values that are not set.
-func (env *envVars) validateAndHydrate() error {
-	if env.yamlFilepath == "" {
-		return fmt.Errorf("%s is not set", yamlFilePathEnv)
-	}
-	if env.port == "" {
-		env.port = defaultPort
-	}
-	return nil
 }
